Enhancing / Updating Univention Zammad-App documentationary (#90)
* added console related information

This will help users that need to use the zammad console to do stuff.

* moved comon issues to an own subpage

and extended it with more cases

* added further information section

This contains a toctree which will lead to subpages with even moaR stuff.

* Introducing App-Settings to Univention-App

This documentary will explain how the app-settings work (and why they might not ;) )

* Update running-console-commands-on-univention.rst

* small text correction

small text correction

* small textual change

small textual change

* small textual changes

small textual changes

* corrected wording

Initially I left too much room which lead to a wrong correction. In the described step we don't use, but try to find files. 

(My bad!)