Style: Downcase all RST directives, fix indentation
The RST convention for directives (e.g., `.. note::`) is to use
lowercase. For reference, see the Sphinx ReStructuredText Primer[0]
or the official ReStructuredText docs.[1]

The convention for indentation in RST appears to be three spaces. This
is not explicitly stated anywhere, but it is visually consistent with
the shape of the directive syntax; i.e.

    .. note:: lorem ipsum dolor
       sit amet consectetur

This commit converts all directives in the given Pull Request[2] to
lowercase, and fixes indentation levels as well. To contain the scope of
changes in this PR, it does not modify directives in other files.

[0]: https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html
[1]: https://docutils.sourceforge.io/docs/ref/rst/directives.html
[2]: https://github.com/zammad/zammad-documentation/pull/123