Clarify visual structure with better headings
This commit fixes multiple problems with the visual organization of the
"Set up Elasticsearch" page:

  * Headings are too verbose
  * Headings are used too much
  * Heading levels do not produce a logical structure

== Headings are too verbose

Headings should be very short. Why?

The purpose of headings is to aid in understanding how each section fits
into the rest of the document. This means that when the reader sees a
heading, she is thinking about not just that heading, but the ones that
come before and after it, as well.

When holding and comparing many different pieces of information, you
want those pieces of information to be as small as possible. Think of it
like walking down a trail and picking up interesting rocks that you find
along the way. If each time you stop to pick up a 2kg rock, soon your
arms will be full, and you will have to either stop walking or put down
all your rocks.

Thus, try "Installation" over "Install Elasticsearch and its Attachment
plugin".

== Headings are used too much

The "Set up Elasticsearch" page contains a lot of information,
and the way it uses headings to organize this information is not
ideal.

For instance, certain portions of the page define *steps in a process:*

  1. Install Elasticsearch and its Attachment plugin
  2. Adjust default settings of Elasticsearch
  3. Configure Zammad to work with Elasticsearch

while others lay out *options to choose from:*

  * Generic install Elasticsearch 5.6, 6.x, 7.x (ingest-attachment)
  * CentOS 7
  * Debian 8
  * Debian 9
  * Ubuntu 16.04 & 18.04

Using the same semantic element (`<h#>`) for two separate organizational
patterns leads to a long document that is difficult to navigate because
everything looks the same.

When someone starts a guide, his goal is to *get to the end.*
But if a reader got halfway through the "Optional settings" section of
this page, it's still not clear how much work is left to go!

This commit does two things to fix this:

  * use field lists[0] for options-to-choose-from, and
  * add a "Step n:" prefix to major headings

== Headings do not produce a logical structure

Like <h#> tags in HTML, ReStructuredText headings form a hierarchy of
information in the document:[1]

  1. # (part)
  2. * (chapter)
  3. = (section)
  4. - (subsection)

Originally, the hierarchy of the "Set up Elasticsearch" page was:

    h1 Set up Elasticsearch
      h2 Install Elasticsearch and its Attachment plugin
      h2 Adjust default settings of Elasticsearch
    h1 Configure Zammad to work with Elasticsearch
    h1 Optional settings for Elasticsearch
    h1 List of fields being stored in Elasticsearch

Part of the problem is that this hierarchy is reflected in the sidebar
of the compiled HTML docs. Multiple h1 headings mean that one page has
multiple top-level entries in the sidebar.

Thus, the structure has been revised to:

    h1 Set up Elasticsearch
      h2 Installation
      h2 Configuration
      h2 Connect Zammad to Elasticsearch
        h3 Optional settings
      h2 Appendix: List of indexed fields

with the Appendix planned for extraction to a separate page in a
subsequent commit.

[0]: https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#field-lists
[1]: https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html#sections