Replace written instructions with "code comments"
In a guide like this, some text gives necessary background information,
but most text is just there to elucidate the code samples.

Use text formatting to help the reader tell the difference at a glance
and avoid reading whole sentences when a short code comment will do.

Instead of:

    === Step 3: Connect Zammad

    First of all we need to tell Zammad where it can find elasticsearch:

        zammad run rails r "Setting.set('es_url', 'http://localhost:9200')"

Try:

    === Step 3: Connect Zammad

        # Set the Elasticsearch server address
        $ zammad run rails r "Setting.set('es_url', 'http://localhost:9200')"

This reduces the mental load of interpreting these instructions.

---

Also, we use the `$` symbol here to indicate an example on the command
line. It has the downside that a user can't just copy-paste multiple
lines into the terminal, so we have left it out of larger code blocks.