Replace :ref: with :doc:
ReStructuredText features different ways of creating internal links:

* `:ref:` allows you to reference any location in the project.
  To use it, you have to add markup at the destination.

* `:doc:` allows you to reference any document in the project.
  It links to the top of the page of the specified document,
  but requires no additional markup.

Virtually every use of `:ref:` was just a link to another page
(document), and not a specific location/section of a page. Thus, there
was lots of unnecessary reference-point markup littered throughout the
project (e.g., a `.. _install-elasticsearch:` line at the top of
`/install/elasticsearch.rst`). This commit removes those lines.

It also removes a largely redundant meta-table-of-contents on the front
page of the docs.