add users email address to console output
This will add the users mail address to the console output. Especially when handling GPDR related deletion requests this is very usefull to see if the user was within the system.

As e.g. Google users tend to use @gmail.com or @googlemail.com you can track these users down easier (because you have a before and after list) to see which users are missing in your reported list.