Increase visual separation in Danger Zone code samples (#125)
* Increase visual separation in Danger Zone code samples

Per comments on 641f241 by Marcel Herrguth:

> You merged two code blocks [in admin/console/dangerzone-for-experts.rst]
> that were divided for a important reason: The user not to run it
> immediately
>
> (I learned in the community that many people have issues reading and
> that sometimes causes bad things)

This commit uses definition lists to clearly separate the steps of
deleting DB records in the Console guides.

* Amend per @mrgeneration's QA

* add users email address to console output

This will add the users mail address to the console output. Especially when handling GPDR related deletion requests this is very usefull to see if the user was within the system.

As e.g. Google users tend to use @gmail.com or @googlemail.com you can track these users down easier (because you have a before and after list) to see which users are missing in your reported list.

Co-authored-by: Marcel <github@thehomeofanime.de>