Add ticket article hints for console (solves issue 99)
This commit solves #99.

It extends the console documentation with Article commands that help users to count note articles as SLA counted articles as well if needed.

It also comes with a small warning for the scope.
It's technically safe to use but changes some functioanlity. :> Thus, it also comes with warnings and hints.