Add ticket article commands for console (SLA communication) (#129)
* Add ticket article hints for console (solves issue 99)

This commit solves #99.

It extends the console documentation with Article commands that help users to count note articles as SLA counted articles as well if needed.

It also comes with a small warning for the scope.
It's technically safe to use but changes some functioanlity. :> Thus, it also comes with warnings and hints.

* Fix wording for community (our -> the) non related

This changes "our Community" to "the Community" which is our preferred wording for the Zammad community forum.

* Update admin/console/working-on-ticket-articles.rst

Co-Authored-By: Ryan Lue <hello@ryanlue.com>

Fixing bad phrases and wording.
Enhancing existing rst code to meet the standards.