Adjust documentation for upcoming 3.4 release (#130)
* Raise documentation to ruby version 2.6.5

This commit provides hints for required ruby versions for Zammad.

* Provide alternative to FTP download

This gives the choice to the user if he wants to get a package from our servers or might rather want to get latest stable from Github.

Github is always more recent, which might be interesting for some users.

* Adjust docker-compose update with a warning for possible changes

Some changes required to docker can be found within the readme of the docker-compose repository. As this repo always holds the most current information, this might be important for users upgrading.

* fix typos

* Add note for software requirenments for source code installations

When updating Zammad on a source code installation, one should always check if all required packages are installed beforehand. The new part will help with the location on where to look.

* Remove big warnings for updates to Zammad 2.9

Hopefully everyone has passed that point already.

* Rework ruby versions and warnings

To have less warning colors on this page I decided to have a slim table to show which ruby versions are required for what Zammad version (below to Zammad 2.5 at the moment) to help users on older versions on their upgrade path.

* Re arrange and extend elasticsearch part of software page

This commit cleans up the Elasticsearch part of our documentation. We'll now provide a clear table telling the supported versions of Elasticsearch for using with Zammad.

Some wordings have been adjusted which should fit alot better than before.

Beside of that, we're finally dropping ES prior 5.5.

* fix typo

* Apply suggestions from Ryan

* enhance wording
* remove white spaces that are not needed
* re-arrange appearance of release urls for better "priotizing"

Co-authored-by: Ryan Lue <hello@ryanlue.com>

* remove lead sentence for ES versions

As suggested by @rlue

* Update prerequisites/software.rst

Provide more detailed information on breaking dependencies and why you'd actually want to have ES running.

Co-authored-by: Ryan Lue <hello@ryanlue.com>

Co-authored-by: Ryan Lue <hello@ryanlue.com>