Update ruby version on software page (#141)
* Update ruby version on software page

With upcoming Zammad 3.5 the ruby version required will change. This pages provides the hint for that.

* specify required versions more detailed

As per hint from Thorsten the ruby change also affects Zammad 3.4.1 and thus we need to be very specific about the versions. :>

* CherryPick typo fix (maintenance)

affected commit was falsely merged into develop

* Fix storage_provider typos

This originally came from @Lauro