Introduce SSO documentation (#147)
* WIP: Introduction of SSO authentication

This is the initial commit for SSO authentication against Kerberos. Highly WIP and not yet ready for the real deal.

* fixed typos

* improve troubleshooting section and code samples

* fix / improve stylings and intendations

* more typos / wordings

* force new lines for environment variable listing

Before:
Endpoint: /auth/sso Accepted Header: X-Forwarded-User Accepted ENV: REMOTE_USER OR HTTP_REMOTE_USER

After:
Endpoint: /auth/sso
Accepted Header: X-Forwarded-User
Accepted ENV: REMOTE_USER OR HTTP_REMOTE_USER


This improves the overview a lot

* Enhance ktutil part

* more back and forth for ktutil

* added hint to add the configuration to the files end

* WIP: Client configuration

* add picture for password prompt

* Provide hint that activation of SSO is required for SSO to work

* provide further possible issues for kerberos sso

* add sso login screencast and how to set internet options on windows

* provide Firefox for Kerberos

* worked on wordings

* improvie firefox screencast

* add support scope limitations as requested by Martin

Martin wanted a specified scope of the topics we can cover. With this we want to ensure to not frustrate people with things we can't provide.

* QA: SSO page

* Add screenshot to SSO page

* QA QA: Proposed changes on SSO page

* More QA QA

Co-authored-by: Ryan Lue <hello@ryanlue.com>