Maintenance: Fix deprecation warning for overwrites & add sphinx-tabs
As we're overwriting some parts of the documentation, the build process was throwing deprecation warnings due to dropped support for a setting.

Also sphinx-tabs allow us to provide information for different e.g. OSes without leaving the same page.

This reduces the navigations complexity and also ensures the page stays as small as possible.